+++ src/game/server.cpp
@@ -3380,11 +3380,11 @@ namespace server
             {
                 std::vector<std::string> prev;
                 explodelist(sv_previousmaps, prev);
-                loopvrev(prev) if(prev[i] == smapname)
+                loopvrev(prev) if (prev[i] == smapname)
                 {
                     prev.erase( prev.begin() + i );
                 }
-                while(prev.size() >= G(maphistory))
+                while (prev.size() >= static_cast<size_t>(G(maphistory)))
                 {
                     prev.pop_back();
                 }
@@ -3395,7 +3395,8 @@ namespace server
                 }
                 prev.clear();
             }
-            if(!buf.empty()) setmods(sv_previousmaps, buf.c_str());
+            if(!buf.empty())
+                setmods(sv_previousmaps, buf.c_str());
         }
         else setmods(sv_previousmaps, "");
